ID-181 Improve navigation wrapping to multiple lines
This fixes a visual bug for primary navigation bars with a large number
of items.  An active item not on the bottom row would be highlighted
with a triangle that appeared to float.

This is visible at the following URL:
  http://www2.warwick.ac.uk/services/universitycards/photos

A class of `nav-wrapped` is applied to navigation bars where any items
have wrapped to a line beneath the first.

When a primary nav wraps, the triangle indicator used to display the
active item is hidden.  The active item still displayed in bold text.
Additionally, the opacity of the text for non-active nav items is
decreased.