ID-188 Fix headroom behaviour on iOS 9 Safari
Use the `transform: translate3d(0, 0, 0);` gambit to make the affixed
portion of the masthead render correctly when scrolling down the page in
Safari.  Behaviour in Chrome on iOS is not as good, but unchanged by
this commit.

Including translate3d in a transform CSS property makes the compiler not
generate prefixed properties for IE and Opera, so the transform property
is duplicated to produce the desired output.