ID-193 Calculate nav affix position depending on title height
When the text in the masthead h1 spans over multiple lines, the nav bar
was obscured as its offset from the top of the page was fixed in CSS.
Now calculating this offset on the client so we can be responsive to
this.

Implemented by writing a style tag to the page and repeating the process
on id7:reflow.  Not sure whether there's a better way of doing this.