ID-195 Allow ID7 search box to appear anywhere
Removes the `.id7-masthead` part from the selector for styling the ID7
search box, to allow it to appear anywhere on the page.

Also removes all mentions of the class `id7-search-box`, which does not
appear to be used anywhere.  The closest is `id7-search-box-container`,
or the ID `id7-search-box`.

Adds the Search component to the components page in the documentation.