ID-196 Fix dropdown strangeness
On xs screen sizes, use the Bootstrap dropdown JS plugin to make
click/tap on a dropdown open the associated submenu.

On larger screens, disable the plugin and use existing CSS to make the
submenu appear on mouseover.

Previously, on non-xs screens, both rules were applied at once which
could lead to the dropdown appearing selected but the submenu not being
visible.