Fix up the feature flag to be more lenient
It appears jQuery's data function converts DOM attributes to e.g.
booleans if they match true/false. The original test assumed the value
would always be a string, and used the strict === comparison operator.

Since this assumption was wrong, it made it impossible to actually enable
the MW feature after PR #45 was merged.