ID-426 Rework how brand colours are calculated.
Implemented WCAG functions for contrast. Moved the definitions of key colours up into pairs of colours that are calculated to pass AA contrast, and updated links and headings, both on the page and over the default box styles, use an alternative colour if the brand colour doesn't work.

Brand testcard now passes contrast for all the new brand colours.

Changed @text-color from #383838 to #202020 for a contrast boost.

Added vitest for LESS unit tests.