ID-414 Update brand colours.
Docs page updated for the colours we currently show (replacing green with teal), and a link to the full brand palette page on Sitebuilder.

The primary purple colour has been changed so that's likely to be the main change noticeable when updating ID7. Otherwise only if you had referenced any of the other brand colours.

May need to wait for ID-426 to push these out as the new colours may not pair with the most contrasting text colour using the existing contrast algorithm.