ID-426 Add a little notice to the top.
To let people know we are working on this.

Also got a bit fed up with the repetitive HTML - seems like somebody didn't know about Liquid templates. Now the brand palette is generated from some YAML data, as is a brand testcard that prints out headings, links and box styles in a range of brand colours.

Adjusted config-options.less so that these brands can be applied at any level in the document, not just to the body. Allows for having multiple themes in one page.