ID-430 adjust breakout, fix mastheads.
Breakout container now uses negative margin instead of a 2D transform which might make it more reliable.

Fixed some masthead stuff by using the correct image, and using a separate image for xs since I hadn't realised the logo on that is actually smaller - on all the other sizes the only other difference is the width of the bleed.

README updates.