ID-443 Remove lodash.
Thanks to modern browsers, we don't need it any more other than for HTML escape, and I've replaced that with a function that does a string replace (which is what lodash did, but in a more roundabout way).

Obviously if your app uses lodash you'll still need it, but you might find it doesn't and you can get rid of a dependency.