ID-437 7.1 is mandatory and painful
Tried to remove redundant styles that 7.1 would just be overriding straight away.

For some reason the masthead went all over the place and the best fix was to replace a bunch of silly floating and margins with a simple flex box, which we can do now because it is not the Victorian era.

Some very old and crufty stuff has gone away at the same time. We don't need to talk about IE9 any more, for example.