SBTWO-12266 Link focus styles.
They were merging with the existing Bootstrap link focus styles, and that seems to be inconsistent about which order they will get merged in, so in some cases the Bootstrap rules were overriding the accessibility ones, and adding a default webkit focus ring colour.

Since this is our copy of Bootstrap we can luckily just clear all that out.