ID-438 Replace logos/marks.
Also a bunch of supporting stuff to make it work.

You can set @id7-gen-logos:2025 without the other features, and things will be adjusted and sized to accomodate the new logo (and also add the crest to the footer, approval pending). However currently the app will need to swap out logo.svg in its template for wordmark.svg as appropriate, so that an app can support both at once until we can chuck out the old one.