ID-454 preventDefault first.
The fix is just to move e.preventDefault() to the start of the method, so that if there's an error in the middle of the handler it doesn't fall through to the default href handling, which if the href starts with `javascript:`, will run the code.

I don't really know why this is more of a concern than any other link someone could add to a page to run Javascript. It's not like anyone can get control over the href of a carousel button more easily than anywhere else.