ID-295 Match cells in a thead.
It's debatable how useful this is as it isn't very readable with all the headings stacked at the top, and I don't know why a presentation (i.e. non-data) table would have header cells. But I've done this fix as it completes what was an only partially implemented attempt at restacking cells on mobile.