ID-465 Fix dependency declarations.
A dependency without a caret says "ID7 must have this exact version" which is annoying when you want to update the version of loglevel or whatever in your app. We should be as loose as possible here to allow it to accept newer versions within the same major version.

Also we had a couple of dev dependencies in the normal dependencies, so apps are pulling `sharp` in through to their node_modules, which should not be happening by default.