ID-467 Flex always.
Allow the mobile view to force the footer to the bottom of the viewport when needed.

All our demo pages were quite long so added a page with only a small amount of content to the demo site to show it in action. Ignored the amount of duplication in the nav generation.