Storybook.
Experiment with adding a web components build with Storybook to document, test and display them in isolation and together. Currently the only stories are for a fake web component button and a Bootstrap panel. Would want to start building the components into the published library, and including the Storybook stories in our published demo site (or even using it to generate most of the examples)